Updating the grunt file to load the test files automatically based on the base test folder locations rather than from a single test file.
Updating the grunt file to load the test files automatically based on
the base test folder locations rather than from a single test file.
Updating the test files to no longer export tests as a function to be
called in the shared test file.
Removing the test file calling each of the individual section test
files.